Compile arr.h as a standalone static library
